backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0
pydantic[email]==2.9.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.27.0
google-auth>=2.45.0
python-multipart==0.0.12
slowapi==0.1.9
aiofiles==24.1.0
pdfminer.six==20231228
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5
sse-starlette==2.1.0
pytest==8.3.0
pytest-asyncio==0.24.0
pyyaml==6.0.2
